django-backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.17
django-bootstrap-v5==1.0.11
django-htmx==1.21.0
djangorestframework==3.15.2
soupsieve==2.6
sqlparse==0.5.3
